V8.0.1945 (#569)
* Update build.gradle

* Update build.gradle

* feat(time spend):

26 minutes

* fix(BUG. no scheduled scan start):

5 minutes

* fix(need long test on windows):

1 hour 14 minutes

* fix:

20 minutes

* feat:

1 hour 57 minutes

* feat:

7 minutes

* feat:

* fix:

* fix:

39 minutes

6 minutes

* fix:

23 seconds

* feat:

* perf(remove redundant):

3 minutes

* fix:

9 minutes

38 seconds

* feat(new pcscan scheduler test):

28 minutes

3 minutes

* fix(BUG):

* fix:

1 hour 7 minutes

15 minutes

* fix:

6 minutes

19 seconds

* fix:

20 minutes

* fix:

* fix:

1 hour 25 minutes

26 seconds

* feat(log from file to db):

39 minutes

* fix:

1 minute

* feat:

1 minute

* fix(BUG): 2 runs

13 minutes

* fix:

20 minutes

1 minute

* fix:

2 hours 54 minutes

* fix(executor):

2 minutes

9 minutes

3 minutes

* fix:

27 minutes

* refactor(getprops):

40 minutes

* fix:

15 minutes

* fix(exec by ThreadConfig):

27 minutes

* fix(sched scan need fix):

22 minutes

* fix(Names Scanner execution):

* fix(mem table props):

* fix:

* fix(by failed tests):

44 minutes

34 seconds

* fix(by TestQuallity):

* fix:

15 minutes

* feat(context hash log):

3 minutes

35 seconds

* perf(remove context get exception):

4 minutes

* perf(Netscanservice):

* perf:

* perf:

* perf:

* perf:

* style:

* test:

13 minutes

+++ modules/networker/src/test/java/ru/vachok/networker/ad/common/ConcreteFolderACLWriterTest.java
@@ -25,7 +25,9 @@ public class ConcreteFolderACLWriterTest {
     @Test
     public void testRun() {
         File fileOwner = new File(currentPath.toAbsolutePath().normalize().toString() + ConstantsFor.FILESYSTEM_SEPARATOR + FileNames.FILENAME_OWNER);
-        Assert.assertTrue(fileOwner.delete());
+        if (fileOwner.exists()) {
+            Assert.assertTrue(fileOwner.delete());
+        }
         ConcreteFolderACLWriter concreteFolderACLWriter = new ConcreteFolderACLWriter(currentPath);
         concreteFolderACLWriter.run();
         Assert.assertTrue(fileOwner.exists());
